Moved BookingCTL tests Added BookingCTL Integration tests for Card, Room and Guest
